src/backend/pyproject.toml
[tool.poetry]
name = "api-drivencoder-com"
version = "0.1.0"
package-mode = false
description = "FastAPI backend for Quant Analytics Platform"
authors = ["Rajkumar Mahato <rajkumarcoder.com@gmail.com>"]
readme = "README.md"

packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

numpy = ">=2.4.0,<3.0.0"
pandas = ">=2.3.3,<3.0.0"

fastapi = ">=0.128.0,<0.129.0"
uvicorn = ">=0.40.0,<0.41.0"

yfinance = ">=1.0,<2.0"
requests = ">=2.32.5,<3.0.0"

redis = ">=7.1.0,<8.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
pydantic-settings = ">=2.12.0,<3.0.0"
lxml = "^6.0.2"
python-jose = "^3.5.0"

[tool.poetry.group.dev.dependencies]
watchdog = ">=6.0.0,<7.0.0"

[tool.poetry.scripts]
start = "starts:start"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
